extraxted xml parser to separate maven module
