+++ core/src/main/java/net/sourceforge/jnlp/JNLPMatcher.java
@@ -37,6 +37,7 @@ exception statement from your version.
 
 package net.sourceforge.jnlp;
 
+import net.adoptopenjdk.icedteaweb.xmlparser.Node;
 import net.sourceforge.jnlp.util.logging.OutputController;
 
 import java.io.InputStream;
