+++ core/src/main/java/net/sourceforge/jnlp/Parser.java
@@ -16,6 +16,8 @@
 // Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 package net.sourceforge.jnlp;
 
+import net.adoptopenjdk.icedteaweb.xmlparser.Node;
+import net.adoptopenjdk.icedteaweb.xmlparser.ParseException;
 import net.sourceforge.jnlp.SecurityDesc.RequestedPermissionLevel;
 import net.sourceforge.jnlp.UpdateDesc.Check;
 import net.sourceforge.jnlp.UpdateDesc.Policy;
@@ -790,7 +792,7 @@ public final class Parser {
     /**
      * @param parent
      * @return the component descriptor.
-     * @throws net.sourceforge.jnlp.ParseException
+     * @throws ParseException
      */
     ComponentDesc getComponent(Node parent) throws ParseException {
 
@@ -825,7 +827,7 @@ public final class Parser {
     /**
      * @return the association descriptor.
      * @param node
-     * @throws net.sourceforge.jnlp.ParseException
+     * @throws ParseException
      */
     private AssociationDesc getAssociation(Node node) throws ParseException {
         String[] extensions = getRequiredAttribute(node, "extensions", null).split(" ");
@@ -1370,8 +1372,8 @@ public final class Parser {
         return result;
     }
 
-    public static final String MALFORMED_PARSER_CLASS = "net.sourceforge.jnlp.MalformedXMLParser";
-    public static final String NORMAL_PARSER_CLASS = "net.sourceforge.jnlp.XMLParser";
+    public static final String MALFORMED_PARSER_CLASS = "net.adoptopenjdk.icedteaweb.xmlparser.MalformedXMLParser";
+    public static final String NORMAL_PARSER_CLASS = "net.adoptopenjdk.icedteaweb.xmlparser.XMLParser";
 
     /**
      * @return the root node from the XML document in the specified input
