+++ core/src/test/java/net/sourceforge/jnlp/JNLPFileTest.java
@@ -37,6 +37,7 @@ exception statement from your version.
 
 package net.sourceforge.jnlp;
 
+import net.adoptopenjdk.icedteaweb.xmlparser.ParseException;
 import net.sourceforge.jnlp.JNLPFile.Match;
 import net.sourceforge.jnlp.annotations.Bug;
 import net.sourceforge.jnlp.mock.MockJNLPFile;
