+++ core/src/test/java/net/sourceforge/jnlp/ParserBasicTest.java
@@ -36,6 +36,8 @@ exception statement from your version.
  */
 package net.sourceforge.jnlp;
 
+import net.adoptopenjdk.icedteaweb.xmlparser.Node;
+import net.adoptopenjdk.icedteaweb.xmlparser.ParseException;
 import net.sourceforge.jnlp.mock.DummyJNLPFile;
 import net.sourceforge.jnlp.util.logging.NoStdOutErrTest;
 import org.junit.Assert;
